requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
cryptography==42.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
